backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
paddleocr==3.4.0
paddlepaddle==3.2.0
transformers>=4.37.0
torch==2.10.0
Pillow==10.2.0
faiss-cpu==1.8.0
sentence-transformers>=2.2.0
pydantic==2.5.3
python-jose==3.3.0
aiofiles==23.2.1
accelerate>=0.26.0
qwen-vl-utils

# 环境变量设置（用于修复 OCR 问题）：
# export FLAGS_enable_pir_api=0
# export FLAGS_enable_onednn=0
# export FLAGS_use_mkldnn=0
# export PADDLE_PDX_DISABLE_MODEL_SOURCE_CHECK=True
